Added multiple functions
Added a fireAt function witch allows both player and tower to shoot.
Added a InRange function which checks wether the player is in range. This is the oncidtion for the FireAt function to be executed.
Added a Projectile Actor class, which both the player and tower will use to spawn and fire